decomposed the code
