+++ frontend/src/components/pages/MainComponent/MyTaskCard/MyTaskCard.module.css
@@ -1,0 +1,18 @@
+.header {
+    margin-bottom: 8px;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.descripiton {
+    display: block;
+    margin-bottom: 8px;
+}
+
+.footer {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-top: 50px;
+}
