+++ frontend/src/components/pages/MainComponent/MyTaskColumn/MyTaskColumnStyles.module.css
@@ -1,3 +1,18 @@
-.column {
+.Column {
     background: var(--card-color);
+}
+
+.columns {
+    display: flex;
+    gap: 16px;
+}
+.column {
+    min-width: 300px;
+    width: 33%;
+}
+
+.noTask {
+    padding: 16px;
+    text-align: center;
+    color: #999;
 }
