+++ frontend/src/components/pages/MainComponent/TaskMeneger.jsx
@@ -19,7 +19,7 @@ const TaskManager = () => {
         { id: "completed", title: "Выполнены", color: "#f6ffed" },
     ];
 
-    const { tasks, setTasks } = useTasks();
+    const { setTasks } = useTasks();
     const [columns] = useState(initialColumns);
     const [isModalVisible, setIsModalVisible] = useState(false);
     const [editingTask, setEditingTask] = useState(null);
@@ -47,29 +47,6 @@ const TaskManager = () => {
         setIsModalVisible(true);
     };
 
-    const getColumnTasks = (columnId) => {
-        return sortTasks(tasks.filter((task) => task.status === columnId));
-    };
-
-    const sortTasks = (tasks) => {
-        return [...tasks].sort((a, b) => {
-            switch (sortOption) {
-                case "title":
-                    return a.title.localeCompare(b.title);
-                case "dueDate":
-                    return dayjs(a.dueDate).diff(dayjs(b.dueDate));
-                case "priority": {
-                    const priorityOrder = { high: 0, medium: 1, low: 2 };
-                    return (
-                        priorityOrder[a.priority] - priorityOrder[b.priority]
-                    );
-                }
-                default:
-                    return 0;
-            }
-        });
-    };
-
     return (
         <Layout style={{ minHeight: "100vh" }}>
             <MyHeader username={username} />
@@ -83,8 +60,8 @@ const TaskManager = () => {
                         />
                         <MyTaskColumn
                             showModal={showModal}
-                            getColumnTasks={getColumnTasks}
                             columns={columns}
+                            sortOption={sortOption}
                         />
                     </Content>
                 </Layout>
