+++ frontend/vite.config.js
@@ -1,8 +1,6 @@
 import react from "@vitejs/plugin-react";
 import { defineConfig } from "vite";
 
-// const VITE_API_URL = "http://localhost:8000/";
-
 export default defineConfig({
     plugins: [react()],
     server: {
